Cleanup, proptypes, factor colors out into single file

+++ server/api/controller/user.controller.js
@@ -5,39 +5,56 @@
 const db = require('../model/model.js');
 
 module.exports = {
-  users: {
-      //create a user to link with trails (creates user instance in users table)
-    post: function (req, res) {
-      db.User.findOrCreate({
+  //create a user to link with trails (creates user instance in users table)
+  post: function (req, res) {
+    db.User.findOrCreate({
+      where: {
+        user: req.body.user
+      }
+    })
+    .then( (user) => {
+      res.sendStatus(201);
+    })
+    .catch( (err) => {
+      res.sendStatus(404);
+    });
+  },
+  //delete user account (removes instance from users table)
+  delete: function(req, res) {
+    db.User.find({
+      where: {
+        user: req.params.id
+      }
+    })
+    .then( (userID) => {
+      if(userID === 0) {
+        console.error('Error DELETE request: user does not exist');
+        res.sendStatus(404);
+        return;
+      }
+      return db.UsersFavorites.destroy({
         where: {
-          user: req.body.user
+          userId: userID.dataValues.id
         }
-      }).then( (user) => {
-        res.sendStatus(201);
       });
-    },
-    //delete user account (removes instance from users table)
-    delete: function(req, res) {
-      db.User.find({
+    })
+    .then( (userfavorite) => {
+      if (userfavorite === 0) {
+        return;
+      }
+      return db.User.destroy({
         where: {
           user: req.params.id
         }
-      }).then( (userID) => {
-        db.UsersFavorites.destroy({
-          where: {
-            userId: userID.dataValues.id
-          }
-        }).then( (userfavorite) => {
-          db.User.destroy({
-            where: {
-              id: userID.dataValues.id
-            }
-          }).then( (user) => {
-            console.log('user destroyed');
-          });
-          res.sendStatus(200);
-        });
       });
-    }
+    })
+    .then( (user) => {
+      console.log('user destroyed');
+      res.sendStatus(200);
+    })
+    .catch( (err) => {
+      console.error('Error deleting user');
+      res.sendStatus(404);
+    });
   }
 };
